+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- Always provide source and target revision for pull request [#23](https://github.com/scm-manager/scm-teamscale-plugin/pull/23)
+- Notify teamscale if pull requests are merged or rejected [#23](https://github.com/scm-manager/scm-teamscale-plugin/pull/23)
+
 ## 1.2.0 - 2020-12-17
 ### Added
 - Mark read only verbs to be able to see teamscale findings in archived repositories ([#8](https://github.com/scm-manager/scm-teamscale-plugin/pull/8))